+++ examples/solid-client-access-grants-demo/src/routes/postAccessRequestForm.ts
@@ -1,5 +1,5 @@
 //
-// Copyright 2022 Inrupt Inc.
+// Copyright Inrupt Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal in
@@ -26,6 +26,7 @@ import {
   getAccessApiEndpoint,
   issueAccessRequest,
   redirectToAccessManagementUi,
+  // eslint-disable-next-line import/no-unresolved
 } from "@inrupt/solid-client-access-grants";
 import { Session } from "@inrupt/solid-client-authn-node";
 import { getEnvironment } from "../utils/getEnvironment";