+++ src/gConsent/discover/getAccessApiEndpoint.ts
@@ -1,5 +1,5 @@
 //
-// Copyright 2022 Inrupt Inc.
+// Copyright Inrupt Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal in
@@ -19,10 +19,10 @@
 // SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-import { UrlString } from "@inrupt/solid-client";
+import type { UrlString } from "@inrupt/solid-client";
 import { fetch as crossFetch } from "@inrupt/universal-fetch";
 import { parse } from "auth-header";
-import { AccessBaseOptions } from "../type/AccessBaseOptions";
+import type { AccessBaseOptions } from "../type/AccessBaseOptions";
 
 async function getAccessEndpointForResource(
   resource: UrlString