+++ src/resource/getSolidDataset.ts
@@ -1,5 +1,5 @@
 //
-// Copyright 2022 Inrupt Inc.
+// Copyright Inrupt Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal in
@@ -19,13 +19,11 @@
 // SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-import {
-  UrlString,
-  getSolidDataset as coreGetSolidDataset,
-} from "@inrupt/solid-client";
-import { VerifiableCredential } from "@inrupt/solid-client-vc";
+import type { UrlString } from "@inrupt/solid-client";
+import { getSolidDataset as coreGetSolidDataset } from "@inrupt/solid-client";
+import type { VerifiableCredential } from "@inrupt/solid-client-vc";
 import { fetchWithVc } from "../fetch";
-import { FetchOptions } from "../type/FetchOptions";
+import type { FetchOptions } from "../type/FetchOptions";
 
 /**
  * Retrieve a Dataset from a Solid Pod using an Access Grant to prove the caller