+++ src/resource/saveSolidDatasetInContainer.test.ts
@@ -1,5 +1,5 @@
 //
-// Copyright 2022 Inrupt Inc.
+// Copyright Inrupt Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal in
@@ -20,7 +20,8 @@
 //
 
 import { it, jest, describe, expect } from "@jest/globals";
-import { mockSolidDatasetFrom, UrlString } from "@inrupt/solid-client";
+import type { UrlString } from "@inrupt/solid-client";
+import { mockSolidDatasetFrom } from "@inrupt/solid-client";
 import { mockAccessRequestVc } from "../gConsent/util/access.mock";
 import { saveSolidDatasetInContainer } from "./saveSolidDatasetInContainer";
 import { fetchWithVc } from "../fetch";