+++ CHANGELOG.md
@@ -4,6 +4,22 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## Unreleased
 
+## [2.3.2](https://github.com/inrupt/solid-client-access-grants-js/releases/tag/v2.3.2) - 2023-06-05
+
+### Deprecation notice
+
+- `resource` marked as deprecated for `getAccessGrantAll`. It can be optionally passed along with `params`.
+
+## [2.3.1](https://github.com/inrupt/solid-client-access-grants-js/releases/tag/v2.3.1) - 2023-05-24
+
+### Documentation
+
+- Buffer API marked as deprecated for `saveFileInContainer` and `overwriteFileInContainer`.
+
+### Bugfixes
+
+- Denied Access Grants now preserve the purpose set in the original Access Request.
+
 ## [2.3.0](https://github.com/inrupt/solid-client-access-grants-js/releases/tag/v2.3.0) - 2023-05-09
 
 ### New feature